requirements.txt
picamera2
pyserial
requests
pybluez
